Fix an issue where the the calendar will skip February if it's printed on the 29th or 30th of any month.
When building a date based on today, you have to make sure to set the date before setting the month. Otherwise, if you change the month and the current date is on a day of the month taht the new month doesn't have, Javascript will be helpful by shifting the whole date to the next month.